+++ CHANGELOG.md
@@ -60,6 +60,7 @@
 * Inventory 2nd pane Actions menu: Adjust New action. Refs UIIN-2439.
 * Add "Local" or "Shared" to flag MARC bib records. Refs UIIN-2522.
 * Edit MARC bib record on Shared Instance. Refs UIIN-2526.
+* Show Effective location facet for all Call Number browse sub-options. Fixes UIIN-2499.
 
 ## [9.4.11](https://github.com/folio-org/ui-inventory/tree/v9.4.11) (2023-08-02)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.4.10...v9.4.11)